# contracts/openapi/sysmind-local-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SysMind AI Local API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/log-analyses": {
      "post": {
        "tags": [
          "log analyses"
        ],
        "summary": "Start Log Analysis",
        "operationId": "start_log_analysis_api_v1_log_analyses_post",
        "parameters": [
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartLogAnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "log analyses"
        ],
        "summary": "Recent Log Analyses",
        "operationId": "recent_log_analyses_api_v1_log_analyses_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogAnalysisListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/log-analyses/{analysis_id}": {
      "get": {
        "tags": [
          "log analyses"
        ],
        "summary": "Get Log Analysis",
        "operationId": "get_log_analysis_api_v1_log_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/log-analyses/{analysis_id}/cancel": {
      "post": {
        "tags": [
          "log analyses"
        ],
        "summary": "Cancel Log Analysis",
        "operationId": "cancel_log_analysis_api_v1_log_analyses__analysis_id__cancel_post",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks": {
      "post": {
        "tags": [
          "agent tasks"
        ],
        "summary": "Start Agent Task",
        "operationId": "start_agent_task_api_v1_tasks_post",
        "parameters": [
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAgentTaskRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "agent tasks"
        ],
        "summary": "Recent Agent Tasks",
        "operationId": "recent_agent_tasks_api_v1_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "agent tasks"
        ],
        "summary": "Get Agent Task",
        "operationId": "get_agent_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/cancel": {
      "post": {
        "tags": [
          "agent tasks"
        ],
        "summary": "Cancel Agent Task",
        "operationId": "cancel_agent_task_api_v1_tasks__task_id__cancel_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/events": {
      "get": {
        "tags": [
          "agent tasks"
        ],
        "summary": "Stream Agent Task Events",
        "operationId": "stream_agent_task_events_api_v1_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentBudgetDto": {
        "properties": {
          "max_rounds": {
            "type": "integer",
            "title": "Max Rounds"
          },
          "max_tool_calls": {
            "type": "integer",
            "title": "Max Tool Calls"
          },
          "timeout_seconds": {
            "type": "number",
            "title": "Timeout Seconds"
          },
          "max_parallel_tools": {
            "type": "integer",
            "title": "Max Parallel Tools"
          }
        },
        "type": "object",
        "required": [
          "max_rounds",
          "max_tool_calls",
          "timeout_seconds",
          "max_parallel_tools"
        ],
        "title": "AgentBudgetDto"
      },
      "AgentBudgetRequest": {
        "properties": {
          "max_rounds": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Max Rounds",
            "default": 4
          },
          "max_tool_calls": {
            "type": "integer",
            "maximum": 16.0,
            "minimum": 0.0,
            "title": "Max Tool Calls",
            "default": 8
          },
          "timeout_seconds": {
            "type": "number",
            "maximum": 120.0,
            "minimum": 2.0,
            "title": "Timeout Seconds",
            "default": 30
          },
          "max_parallel_tools": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Max Parallel Tools",
            "default": 2
          }
        },
        "type": "object",
        "title": "AgentBudgetRequest"
      },
      "AgentTaskListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentTaskResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "AgentTaskListResponse"
      },
      "AgentTaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "created",
              "planning",
              "running_tools",
              "analyzing",
              "waiting_user_input",
              "completed",
              "cancelling",
              "cancelled",
              "failed",
              "timed_out",
              "interrupted"
            ],
            "title": "Status"
          },
          "user_goal": {
            "type": "string",
            "title": "User Goal"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "allowed_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Tools"
          },
          "budget": {
            "$ref": "#/components/schemas/AgentBudgetDto"
          },
          "current_round": {
            "type": "integer",
            "title": "Current Round"
          },
          "tool_call_count": {
            "type": "integer",
            "title": "Tool Call Count"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "working_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Working Summary"
          },
          "final_output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Output"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "failure_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Message"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/AgentToolCallDto"
            },
            "type": "array",
            "title": "Tool Calls"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "user_goal",
          "provider",
          "allowed_tools",
          "budget",
          "current_round",
          "tool_call_count",
          "progress",
          "working_summary",
          "final_output",
          "failure_code",
          "failure_message",
          "cancel_requested",
          "created_at",
          "started_at",
          "finished_at",
          "schema_version"
        ],
        "title": "AgentTaskResponse"
      },
      "AgentToolCallDto": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "provider_call_id": {
            "type": "string",
            "title": "Provider Call Id"
          },
          "tool_name": {
            "type": "string",
            "title": "Tool Name"
          },
          "tool_version": {
            "type": "string",
            "title": "Tool Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "arguments_hash": {
            "type": "string",
            "title": "Arguments Hash"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "result_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Summary"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "id",
          "provider_call_id",
          "tool_name",
          "tool_version",
          "status",
          "arguments_hash",
          "risk_level",
          "started_at",
          "finished_at",
          "duration_ms",
          "result_summary",
          "error_code",
          "error_message"
        ],
        "title": "AgentToolCallDto"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogAnalysisFailureDto": {
        "properties": {
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "code",
          "message"
        ],
        "title": "LogAnalysisFailureDto"
      },
      "LogAnalysisListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LogAnalysisResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "LogAnalysisListResponse"
      },
      "LogAnalysisQueryDto": {
        "properties": {
          "channels": {
            "items": {
              "type": "string",
              "enum": [
                "Application",
                "System"
              ]
            },
            "type": "array",
            "title": "Channels"
          },
          "lookback_hours": {
            "type": "integer",
            "title": "Lookback Hours"
          },
          "levels": {
            "items": {
              "type": "string",
              "enum": [
                "critical",
                "error",
                "warning",
                "information"
              ]
            },
            "type": "array",
            "title": "Levels"
          },
          "event_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Event Ids"
          },
          "max_events": {
            "type": "integer",
            "title": "Max Events"
          }
        },
        "type": "object",
        "required": [
          "channels",
          "lookback_hours",
          "levels",
          "event_ids",
          "max_events"
        ],
        "title": "LogAnalysisQueryDto"
      },
      "LogAnalysisResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "partial",
              "cancelled",
              "failed"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "query": {
            "$ref": "#/components/schemas/LogAnalysisQueryDto"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LogAnalysisSummaryDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "failures": {
            "items": {
              "$ref": "#/components/schemas/LogAnalysisFailureDto"
            },
            "type": "array",
            "title": "Failures"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "progress",
          "current_step",
          "started_at",
          "finished_at",
          "query",
          "summary",
          "failures",
          "schema_version"
        ],
        "title": "LogAnalysisResponse"
      },
      "LogAnalysisSummaryDto": {
        "properties": {
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventEvidenceDto"
            },
            "type": "array",
            "title": "Events"
          },
          "event_groups": {
            "items": {
              "$ref": "#/components/schemas/EventGroupDto"
            },
            "type": "array",
            "title": "Event Groups"
          },
          "crash_groups": {
            "items": {
              "$ref": "#/components/schemas/CrashGroupDto"
            },
            "type": "array",
            "title": "Crash Groups"
          },
          "notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notice"
          }
        },
        "type": "object",
        "required": [
          "event_count",
          "events",
          "event_groups",
          "crash_groups"
        ],
        "title": "LogAnalysisSummaryDto"
      },
...
      "StartAgentTaskRequest": {
        "properties": {
          "user_goal": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "User Goal"
          },
          "allowed_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Allowed Tools"
          },
          "budget": {
            "$ref": "#/components/schemas/AgentBudgetRequest"
          }
        },
        "type": "object",
        "required": [
          "user_goal"
        ],
        "title": "StartAgentTaskRequest"
      },
...
      "StartLogAnalysisRequest": {
        "properties": {
          "channels": {
            "items": {
              "type": "string",
              "enum": [
                "Application",
                "System"
              ]
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 1,
            "title": "Channels",
            "default": [
              "Application",
              "System"
            ]
          },
          "lookback_hours": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Lookback Hours",
            "default": 24
          },
          "levels": {
            "items": {
              "type": "string",
              "enum": [
                "critical",
                "error",
                "warning",
                "information"
              ]
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 1,
            "title": "Levels",
            "default": [
              "critical",
              "error",
              "warning"
            ]
          },
          "event_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Event Ids"
          },
          "max_events": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Max Events",
            "default": 100
          }
        },
        "type": "object",
        "title": "StartLogAnalysisRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}